requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
certifi==2023.5.7